pyproject.toml
[tool.poetry]
name = "fast-fuzzy-matching"
version = "0.1.0"
description = "Perform fast approximate string matching."
license = "MIT"
authors = ["Tom Matthews <tomukmatthews@gmail.com>"]
readme = "README.md"
packages = [{include = "fast_fuzzy_matching"}]
# packages = [{include = "lcs", from = "fast_fuzzy_matching"}]
# packages = [
#     { include = "string_matching", from = "app/utils" },
# ]


[tool.poetry.dependencies]
python = "^3.10"
sparse-dot-topn = "^0.3.3"
pandas = "^1.5.2"
scipy = "^1.9.3"
pybind11 = "^2.10.1"
scikit-learn = "^1.1.3"
setuptools = "^65.6.3"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
pytest = "^7.2.0"
pytest-icdiff = "^0.6"
pylcs = "^0.0.7"

[tool.black]
line-length = 120
skip-string-normalization = 1 

[tool.isort]
line_length = 120
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.build]
script = "build.py"
